feat: add logics toggle for product cards
